Separate WaspFile into modules
